Considerable progress towards sorting script
Initial creation of, and lots of progress towards, the sorting script.  Should be mentioned that these scripts will eventually be merged into the C# program itself to remove the need for additional files outside of the executeable (should the ports to C# prove themselves performance-wise). Also added some minor documentation to delete.bat and dummyfiles.bat